[TST/MAINT] move misplaced results_quantile_regression to results dir
