+++ statsmodels/regression/tests/results/results_quantile_regression.py
